+++ changelog.md
@@ -8,6 +8,7 @@
 - Update to project template 4.3.3
 - Removed submodules and added the libraries as packages
 - Updated question flow adding social closeness input in the question creation flow instead of similarity and reason
+- Updated app commands: renamed `/question` into `/ask` and `/answer` into `/questions`.
 
 ## 1.0.10
 - Changed the `project` field in the messages logged into the Wenet platform. Now the `appId` is used, instead of the `PROJECT_NAME` environment variable